AB#4281: persist rollup bucket/watermark ms as Int64 in the Mongo store
RtRollupArchive.BucketSizeMs / WatermarkLagMs are Int64 as of
System.StreamData 1.6.3. MongoRollupArchiveRuntimeStore.InsertAsync cast
bucketSize.TotalMilliseconds / watermarkLag.TotalMilliseconds to (int) —
which truncates a calendar-month/year width before the widening even
applies. Cast to (long).

Also adds AB#4282 regression guards to RecomputeBucketEnumeratorTests:
calendar-month over 3 years (36 buckets), calendar-year over a multi-year
range, and ISO-8601-week over 3 years (157 buckets) all enumerate via pure
DateTime calendar arithmetic with a >Int32 BucketSize and never overflow.

+++ src/Runtime.Engine.MongoDb/StreamData/MongoRollupArchiveRuntimeStore.cs
@@ -81,8 +81,11 @@ public sealed class MongoRollupArchiveRuntimeStore : IRollupArchiveRuntimeStore
             TargetCkTypeId = targetCkTypeId.ToString(),
             Status = RtCkArchiveStatusEnum.Created,
             SourceArchiveRtId = sourceArchiveRtId.ToString(),
-            BucketSizeMs = (int)bucketSize.TotalMilliseconds,
-            WatermarkLagMs = (int)watermarkLag.TotalMilliseconds,
+            // AB#4281: BucketSizeMs / WatermarkLagMs are Int64 (System.StreamData 1.6.3). Cast to long
+            // — a calendar-month (2,419,200,000 ms) or calendar-year (31,536,000,000 ms) bucket width
+            // exceeds Int32.MaxValue, so an (int) cast would silently truncate before the widening.
+            BucketSizeMs = (long)bucketSize.TotalMilliseconds,
+            WatermarkLagMs = (long)watermarkLag.TotalMilliseconds,
             // LastAggregatedBucketEnd starts null — set by ArchiveLifecycleService.ActivateAsync to
             // the activation timestamp (truncated to the bucket boundary). The orchestrator then
             // advances it forward, never backwards (except via rewindRollupWatermark). Concept §5.
